Update run.js
url for installing kubectl receives the semantic version string like "1.15.0" and it adds the letter 'v' before it. To not change the documentation, I believe this edit would solve it. 

What solved for me was was to use it like this:
```
- uses: azure/setup-kubectl@v1
        with:
          version: 'v1.15.0'
        id: install
```

But then I was not able to run kubectl commands... :/